// src/Gatari/Interfaces/IGatariBeatmapCompact.ts
export interface IGatariBeatmapCompact {
  ar: number;
  artist: string;
  beatmap_id: number;
  beatmap_md5: string;
  beatmapset_id: number;
  bpm: number;
  creator: string;
  difficulty: number;
  fc: number;
  hit_length: number;
  od: number;
  ranked: number;
  ranked_status_frozen: number;
  song_name: string;
  title: string;
  version: string;
}